change decimal to double

+++ USASales/USASales/Models/Product.cs
@@ -5,7 +5,7 @@
         public long Id { get; set; }
         public string Name { get; set; }
         public string Category { get; set; }
-        public decimal WholesalePrice { get; set; }
-        public decimal GrossPrice { get; set; }
+        public double WholesalePrice { get; set; }
+        public double GrossPrice { get; set; }
     }
 }
